Personal Medical Cost Project
This is an analysis of the data set "Medical Cost Personal Datasets". The goal of this project is to find which of the predictors have the most impact on the prediction of the personal medical cost. I am interested in how does the different factor influence on the individual medical cost billed by health insurance, such as gender, age and so on.